outsources search field options
