+++ app/src/main/res/layout/customize_app_drawer_fragment.xml
@@ -11,73 +11,28 @@
         android:id="@+id/customize_app_drawer_fragment_visible_apps"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/_16sdp"
+        android:layout_marginLeft="@dimen/_16sdp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="@dimen/_16sdp"
+        android:layout_marginRight="@dimen/_16sdp"
+        android:text="@string/customize_app_drawer_fragment_visible_apps"
         android:textAppearance="@style/TextAppearance.AppCompat"
-        android:layout_marginStart="@dimen/_16sdp"
-        android:layout_marginLeft="@dimen/_16sdp"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="@dimen/_16sdp"
-        android:layout_marginRight="@dimen/_16sdp"
         android:textSize="@dimen/_20ssp"
-        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        android:text="@string/customize_app_drawer_fragment_visible_apps" />
-
-    <View
-        android:id="@+id/divider"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/_1sdp"
-        android:layout_marginTop="32dp"
-        android:background="@color/primaryColorDark"
-        android:layout_marginStart="@dimen/_16sdp"
-        android:layout_marginLeft="@dimen/_16sdp"
-        android:layout_marginEnd="@dimen/_16sdp"
-        android:layout_marginRight="@dimen/_16sdp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/customize_app_drawer_fragment_visible_apps" />
-
-    <androidx.appcompat.widget.SwitchCompat
-        android:id="@+id/customize_app_drawer_fragment_search_bar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textAppearance="@style/TextAppearance.AppCompat"
-        android:layout_marginStart="@dimen/_16sdp"
-        android:layout_marginLeft="@dimen/_16sdp"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="@dimen/_16sdp"
-        android:layout_marginRight="@dimen/_16sdp"
-        android:textSize="@dimen/_20ssp"
-        app:layout_constraintTop_toBottomOf="@id/divider"
-        app:layout_constraintStart_toStartOf="parent"
-        android:text="@string/customize_app_drawer_fragment_show_search_bar" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/customize_app_drawer_fragment_search_bar_position"
-        android:layout_width="match_parent"
+        android:id="@+id/customize_app_drawer_fragment_search_options"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textAppearance="@style/TextAppearance.AppCompat"
-        android:layout_marginStart="@dimen/_16sdp"
-        android:layout_marginLeft="@dimen/_16sdp"
         android:layout_marginTop="32dp"
         android:layout_marginEnd="@dimen/_16sdp"
-        android:layout_marginRight="@dimen/_16sdp"
+        android:text="@string/customize_app_drawer_fragment_search_field_options"
+        android:textAppearance="@style/TextAppearance.AppCompat"
         android:textSize="@dimen/_20ssp"
-        app:layout_constraintTop_toBottomOf="@id/customize_app_drawer_fragment_search_bar"
-        app:layout_constraintStart_toStartOf="parent"
-        android:text="@string/customize_app_drawer_fragment_search_bar_position" />
+        app:layout_constraintStart_toStartOf="@id/customize_app_drawer_fragment_visible_apps"
+        app:layout_constraintTop_toBottomOf="@id/customize_app_drawer_fragment_visible_apps" />
 
-    <androidx.appcompat.widget.SwitchCompat
-        android:id="@+id/customize_app_drawer_open_keyboard_switch"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/_16sdp"
-        android:layout_marginLeft="@dimen/_16sdp"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="@dimen/_16sdp"
-        android:layout_marginRight="@dimen/_16sdp"
-        android:layout_marginBottom="32dp"
-        android:text="@string/customize_app_drawer_fragment_open_keyboard"
-        android:textAppearance="@style/TextAppearance.AppCompat"
-        android:textSize="@dimen/_20ssp"
-        app:layout_constraintTop_toBottomOf="@id/customize_app_drawer_fragment_search_bar_position"
-        app:layout_constraintStart_toStartOf="parent"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
