+++ app/src/main/res/values/strings.xml
@@ -47,8 +47,9 @@
     <string name="customize_app_drawer_fragment_auto_theme_wallpaper_text">Set theme background as wallpaper</string>
     <string name="customize_app_drawer_fragment_auto_theme_wallpaper_subtext_no_default_launcher">App needs to be default launcher</string>
     <string name="customize_app_drawer_fragment_visible_apps">Visible Apps</string>
-    <string name="customize_app_drawer_fragment_show_search_bar">Show Search Bar</string>
-    <string name="customize_app_drawer_fragment_search_bar_position">Search Bar Position</string>
+    <string name="customize_app_drawer_fragment_search_field_options">Search Field Options</string>
+    <string name="customize_app_drawer_fragment_show_search_bar">Show Search Field</string>
+    <string name="customize_app_drawer_fragment_search_bar_position">Search Field Position</string>
     <string-array name="search_bar_position_array">
         <item>Top</item>
         <item>Bottom</item>
@@ -64,5 +65,6 @@
     <string name="remove_all_apps_dialog_title" translatable="false">Remove All Apps</string>
     <string name="remove_all_apps_dialog_message" translatable="false">"This action will not uninstall your apps."
     "It is just to confirm if you're clearing this list on purpose"</string>
+
 
 </resources>
